+++ resources/lang/de/validation.php
@@ -206,5 +206,8 @@ return [
         'lockSettingsHideUserList'          => 'Teilnehmerliste verbergen',
         'role'                              => 'Rolle',
         'file'                              => 'Datei',
+        'permissions'                       => 'Rechte',
+        'updated_at'                        => 'Aktualisiert am',
+        'room_limit'                        => 'Max. Anzahl an Räumen'
     ],
 ];