+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Authentication with LDAP and Email-Address ([#1], [#3])
 - A loading screen for loading events of the application ([#6], [#11])
 - Localization with a locale switcher and saving selected locale for authenticated users ([#8], [#13])
-- Flash messages on a successful login and on errors catched by the global vue error handler ([#7], [#16])
+- Flash messages on a successful login and on errors catched by the global vue error handler ([#7], [#16],[#44],[#43])
 - Added Footer with Impress and Privacy policy ([#9], [#17])
 - Added authenticator type to users model ([#19], [#21])
 - Roles and permissions concept ([#12], [#22])
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added file upload for room owner and file download for participants ([#4], [#18])
 - Added global settings, .env as default; added logo path and room limit ([#34],[#36])
 - Added setting menu for administrators ([#35], [#38])
+- Added a middleware to check whether the request is in sync with model of the database and not stale ([#40], [#41])
 
 [#1]: https://github.com/THM-Health/PILOS/issues/1
 [#3]: https://github.com/THM-Health/PILOS/pull/3
@@ -40,5 +41,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#35]: https://github.com/THM-Health/PILOS/pull/35
 [#36]: https://github.com/THM-Health/PILOS/pull/36
 [#38]: https://github.com/THM-Health/PILOS/issues/38
+[#40]: https://github.com/THM-Health/PILOS/issues/40
+[#41]: https://github.com/THM-Health/PILOS/pull/41
+[#43]: https://github.com/THM-Health/PILOS/issues/43
+[#44]: https://github.com/THM-Health/PILOS/pull/44
 
 [unreleased]: https://github.com/THM-Health/PILOS/compare/3c8359cdb0395546fe97aeabf1a40f93002b182c...HEAD